+++ src/constants/index.jsx
@@ -1,0 +1,16 @@
+export const currency = [
+  { currency: "USD", name: "미국 USD" },
+  { currency: "JPY", name: "일본 JPY 100" },
+  { currency: "EUR", name: "유럽연합 EUR" },
+  { currency: "CNY", name: "중국 CNY" },
+  { currency: "GBP", name: "영국 GBP" },
+  { currency: "AUD", name: "호주 AUD" },
+  { currency: "CAD", name: "캐나다 CAD" },
+  { currency: "NZD", name: "뉴질랜드 NZD" },
+  { currency: "TWD", name: "대만 TWD" },
+  { currency: "HKD", name: "홍콩 HKD" },
+  { currency: "THB", name: "태국 THB" },
+  { currency: "PHP", name: "필리핀 PHP" },
+  { currency: "RUB", name: "러시아 RUB" },
+  { currency: "MOP", name: "마카오 MOP" },
+];
